JOC-1767 - numOfFilesFrom,numOfFilesTo - changed to Long instead of Integer (same type as for numOfFiles - transfer)

+++ sos-webservices/sos-webservices-joc-global/src/main/java/com/sos/joc/db/yade/JocYadeFilter.java
@@ -29,8 +29,8 @@ public class JocYadeFilter {
     private Set<String> profiles;
     private Set<String> workflowNames;
     private Integer limit;
-    private Integer numOfFilesFrom;
-    private Integer numOfFilesTo;
+    private Long numOfFilesFrom;
+    private Long numOfFilesTo;
     private Date dateFrom;
     private Date dateTo;
 
@@ -195,19 +195,19 @@ public class JocYadeFilter {
         limit = val;
     }
 
-    public Integer getNumOfFilesFrom() {
+    public Long getNumOfFilesFrom() {
         return numOfFilesFrom;
     }
 
-    public void setNumOfFilesFrom(Integer val) {
+    public void setNumOfFilesFrom(Long val) {
         numOfFilesFrom = val;
     }
 
-    public Integer getNumOfFilesTo() {
+    public Long getNumOfFilesTo() {
         return numOfFilesTo;
     }
 
-    public void setNumOfFilesTo(Integer val) {
+    public void setNumOfFilesTo(Long val) {
         numOfFilesTo = val;
     }
